Support user snippets
https://github.com/bluefeet/GitLab-API-v3/issues/25

GitLab 8.15 introduced the concept of personal/user (non-project)
snippets.

http://docs.gitlab.com/ce/user/snippets.html

https://gitlab.com/gitlab-org/gitlab-ce/blob/master/doc/api/snippets.md

This commit adds config and methods for the new snippet API operations.

+ Add `sections/snippets.yml` file for user snippets.

+ Add `snippets` section to the `author/config.yml` file.

+ Rename the project snippet documentation section in the
`author/config.yml` file to reduce ambiguity.

+ Regenerate v3.pm with user snippets methods and docs:

    perl generate.pl > ../lib/GitLab/API/v3.pm

Tested via `bin/gitlab-api-v3` to make sure that the new user snippet
methods worked and that the old project snippet methods still worked
also.

NOTE: I was unable to test via `t/branches.t` as that failed with:

    skipped: Could not find a private project owned by the tokened user.

Both on master, and on my new branch. My gitlab account does have a
private project set up, and I was able to use the `bin/gitlab-api-v3
projects` to see it and verify that it was set to public=0.