First shot at ditching Role::REST::Client.
The problem with Role::REST::Client is it goes through so many hoops
that it is often impossible (without submitting a pull request) to
change its behavior.  In this case the issue was that the allow_nonref
option needed to be set on the JSON encoder.

Also, our target is the GitLab REST API, not *any* REST API.  This new
REST logic is not all that much code to maintane because it is
targetting the exact design and intricacies of the GitLab REST API,
nothing else.

By custom writing the REST logic a lot more can happen and things like
properly supporting file uploads can finally be implemented.

Definately some refactoring to go, but I think this first implementation
is sound.

Also, this was all inspired by the previous "NS" (no serializer) pull
request which, on closer inspection, turned out to be the incorrect
solution.  The correct solution is to set the allow_nonref flag and
continue parsing the response body as JSON, which is what this commit
does.

Most of the NS stuff has been reverted as part of this commit.  The one
bit kept is the added tests.  Thank you royce55 for the inspiration!