Return the raw response content when decode is false.
Returning the HTTP::Tiny response breaks the contract this API tries to
adhere to; that the HTTP part of things be transparent and hidden away.

If useful, in a future commit, I coould see adding a way to get at the
underlying HTTP::Tiny request/response metadata, but that would probably
be more like a debug option.